+++ test/components/strategy/index.dart
@@ -1,0 +1,2 @@
+export 'strategy.controller.dart';
+export 'strategy.model.dart';